Now all the inserts in DB are unique
Made sure the Customer is check before added again, the Address, the Shipping_Detail.
Finalized the WebFlow to permit to update or have the same Shipping as Billing
Fixed a bug that was messing up the Billing address based on Shipping addres update (I modified the setter for the ShippingDetail setShippingAddress.